+++ test/record/user_adapter_test.rb
@@ -1,0 +1,29 @@
+require 'test_helper'
+
+class Gecko::Record::UserAdapterTest < Minitest::Test
+  include TestingAdapter
+  include SharedAdapterExamples
+
+  let(:adapter)       { @client.User }
+  let(:plural_name)   { 'users' }
+  let(:record_class)  { Gecko::Record::User }
+
+  def test_initializes_adapter
+    assert_instance_of(Gecko::Record::UserAdapter, @client.User)
+  end
+
+  undef :test_adapter_count
+
+  # Can't build users via API
+  undef :test_build
+  undef :test_build_with_attributes
+  undef :test_saving_new_record
+  undef :test_saving_new_invalid_record
+
+  def test_current
+    VCR.use_cassette('users#current') do
+      assert_instance_of(Gecko::Record::User, @client.User.current)
+      assert(@client.User.current, 'User.current is identity mapped')
+    end
+  end
+end